requirements.txt
pandas
numpy
python-docx
rank-bm25
transformers
torch
gradio